+++ src/toker-express.ts
@@ -1,6 +1,7 @@
 
 export {
   refresh,
+  parseBearerToken,
   decodeAccess,
   decodeRefresh,
 } from './jwt';